+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 ### Fixes & Improvements
 
 - helper [objectToFormData](src/helpers/objectToFormData.ts). Added option bracket-notation (use for NodeJS) and dot-notation (use for .NET)
+- [PopupElement](src/popup/popupElement.ts). Fixed _rendered out of viewport on custom fitElement and scrollable parent_
+
+---
 
 ## 1.2.3 (Mar 5, 2025)